fixed scroll effect bug
I added the logic to go to the selected section in the navbar menu. The Moon exhibited strange behavior: when outside of Main, if I clicked on Main and had scrolled the mouse down, it wouldn’t reappear; it would only come back when I scrolled the mouse to the top of the page. And if I was in Main and opened the menu to click, for example, on Contact, the Moon didn’t do its animation and stayed in the background.